start breaking out and documenting stats apis (WIP)
