+++ package.json
@@ -68,8 +68,8 @@
     "viewer:testuserdebug": "(cd viewer && nodemon viewer.js -c ../tests/config.test.ini -n testuser --debug) & webpack -d --no-stats --watch --progress --config viewer/vueapp/build/webpack.dev.conf.js",
     "viewer:multiviewer": "cd viewer && node multies.js -c ../tests/config.test.ini -n all",
     "viewer:testmultiviewer": "(cd viewer && nodemon viewer.js -c ../tests/config.test.ini -n testmulties) & webpack -d --no-stats --watch --progress --config viewer/vueapp/build/webpack.dev.conf.js",
-    "viewer:doc": "jsdoc2md --partial viewer/jsdoc/*.hbs --helper viewer/jsdoc/dochelpers.js --files viewer/apiSessions.js viewer/apiConnections.js viewer/viewer.js > ../arkimeweb/_wiki/api_docs.md",
-    "viewer:doc-location": "jsdoc2md --partial viewer/jsdoc/*.hbs --helper viewer/jsdoc/dochelpers.js --files viewer/apiSessions.js viewer/apiConnections.js viewer/viewer.js >",
+    "viewer:doc": "jsdoc2md --partial viewer/jsdoc/*.hbs --helper viewer/jsdoc/dochelpers.js --files viewer/*.js > ../arkimeweb/_wiki/api_docs.md",
+    "viewer:doc-location": "jsdoc2md --partial viewer/jsdoc/*.hbs --helper viewer/jsdoc/dochelpers.js --files viewer/*.js >",
     "viewer:lint": "eslint --ext .js,.vue viewer capture/plugins",
     "viewer:lint-fix": "eslint --fix --ext .js,.vue viewer",
     "parliament:build": "cd parliament && node vueapp/build/build.js",
