+++ viewer/vueapp/src/components/stats/EsIndices.vue
@@ -178,7 +178,7 @@ export default {
       this.$emit('confirm', `Delete ${indexName}`, indexName);
     },
     deleteIndex (indexName) {
-      this.$http.delete(`esindices/${indexName}`)
+      this.$http.delete(`api/esindices/${indexName}`)
         .then((response) => {
           for (let i = 0; i < this.stats.length; i++) {
             if (this.stats[i].index === indexName) {
@@ -191,21 +191,21 @@ export default {
         });
     },
     optimizeIndex (indexName) {
-      this.$http.post(`esindices/${indexName}/optimize`)
+      this.$http.post(`api/esindices/${indexName}/optimize`)
         .then((response) => {
         }, (error) => {
           this.$emit('errored', error.text || error);
         });
     },
     closeIndex (indexName) {
-      this.$http.post(`esindices/${indexName}/close`)
+      this.$http.post(`api/esindices/${indexName}/close`)
         .then((response) => {
         }, (error) => {
           this.$emit('errored', error.text || error);
         });
     },
     openIndex (indexName) {
-      this.$http.post(`esindices/${indexName}/open`)
+      this.$http.post(`api/esindices/${indexName}/open`)
         .then((response) => {
         }, (error) => {
           this.$emit('errored', error.text || error);
@@ -231,7 +231,7 @@ export default {
       if (desc !== undefined) { this.query.desc = desc; }
       if (sortField) { this.query.sortField = sortField; }
 
-      this.$http.get('esindices/list', { params: this.query })
+      this.$http.get('api/esindices', { params: this.query })
         .then((response) => {
           respondedAt = Date.now();
           this.error = '';
